+++ backend/src/main/kotlin/com/davismariotti/campalert/service/sms/AwaitingContext.kt
@@ -2,5 +2,5 @@ package com.davismariotti.campalert.service.sms
 
 data class AwaitingContext(
     val intent: String,
-    val requestIds: List<Int>,
+    val requestIds: List<Long>,
 )
